demo/requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
python-jose
pydantic
email-validator
jinja2
pytest
httpx
